Refactor validation/constraints
As of this commit, the constraints feature is used for built-in
validation, such as checking that numbers are integers (e.g. in
IntegerPreference).

This makes validate (previously isValidValue) play a less prominent
role: It is only necessary to define when values cannot be validated by
themselves, for example in MultichoicePreference, where it must be
checked that a value is among the available options.

Also, stricter constraints have been placed on NumericPreference values:
NaN and (-)Infinity are no longer valid.