Improve fromInvalid (now toValid) structure
fromInvalid is not a good name, because the function does not (and is
not supposed to) work only for invalid values. Its purpose is just to
always give back a valid one, and it can do so even if its input is
already valid.

Ideally, toValid should be implemented as the identity function for
valid values. For invalid values, it should try to produce a sensible
valid one. For example, the RangePreference implementation caps values
to the interval [min, max].